More inspection fixes

+++ jgrapht-core/src/main/java/org/jgrapht/experimental/alg/color/GreedyColoring.java
@@ -89,7 +89,7 @@ public class GreedyColoring<V, E>
         for (int i = 0; i < _neighbors.length; i++) {
             while (!buckets.get(i).isEmpty()) {
                 final int s = buckets.get(i).size() - 1;
-                final int vertex = (Integer) buckets.get(i).get(s);
+                final int vertex = buckets.get(i).get(s);
                 buckets.get(i).remove(s);
                 degree[vertex] = -1;
                 order[index--] = vertex;
@@ -135,7 +135,7 @@ public class GreedyColoring<V, E>
             satur[v] = -1;
             index++;
             for (int j = 0; j < _neighbors[v].length; j++) {
-                final int nb = (int) _neighbors[v][j];
+                final int nb = _neighbors[v][j];
                 final int bi = bucketIndex[nb];
                 if (satur[nb] >= 0) {
                     if (bi != cumBucketSize[satur[nb]] - 1) {
